*** empty log message ***
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@10422 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/GC/aoa.c
@@ -364,7 +364,7 @@ Object *AOAallocate(long numbytes)
 			       / (totalAOASize / 100)) 
 	    > (unsigned long)AOAPercentage)) {
 #else
-    if (totalAOASize > AOAMinSizeForGC) {
+    if ((unsigned long)totalAOASize > AOAMinSizeForGC) {
 #endif
 
       OverFullFlag = 2;
@@ -1566,11 +1566,11 @@ void scanOrigins(Object *theObj, void (*originAction)(Object **theCell))
 void SetAOANeedCompaction(void)
 {
   if (OverFullFlag) {
-    if (totalAOASize > AOAMinSizeForGC) {
+    if ((unsigned long)totalAOASize > AOAMinSizeForGC) {
       AOANeedCompaction = TRUE;
     }
   } else {
-    if (totalFree < AOAFreeAtGC/2) {
+    if ((unsigned long)totalFree < AOAFreeAtGC/2) {
       AOANeedCompaction = TRUE;
     }
   }
